provide userId moved to core module
