+++ auth/build.gradle.kts
@@ -33,10 +33,10 @@ android {
 }
 
 dependencies {
-
-//    implementation(project(":core"))
+    
+    implementation(project(":core"))
     implementation(Dependencies.appcompat)
-
+    
     implementation(platform(Dependencies.firebaseBom))
     implementation(Dependencies.firebaseAuth)
     implementation(Dependencies.googlePlayAuth)
