+++ core/src/main/java/com/zhigaras/core/ProvideUserId.kt
@@ -1,0 +1,7 @@
+package com.zhigaras.core
+
+interface ProvideUserId {
+    
+    fun provide(): String
+    
+}
